third_party/blink/web_tests/editing/selection/focus-and-display-none-and-redisplay-expected.txt

Unfocused TEXTAREA should accept editing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is not textarea
PASS textarea.value is "addoriginal"
PASS successfullyParsed is true

TEST COMPLETE
